requirements.txt
flask
flask-sqlalchemy
flask-migrate
gunicorn
psycopg2
flask-cors
flask-security
flask-markdown
elasticsearch
lxml.html
requests
